requirements.txt
# Core dependencies
torch>=1.13.0
torchvision>=0.14.0
torchaudio>=0.13.0
transformers>=4.25.0
datasets>=2.8.0
tokenizers>=0.13.0
accelerate>=0.15.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0
Pillow>=9.0.0
opencv-python>=4.6.0
librosa>=0.9.0
PyPDF2>=3.0.0
python-docx>=0.8.0

# Configuration and utilities
pyyaml>=6.0
pydantic>=1.10.0
psutil>=5.9.0
python-dotenv>=0.19.0

# Web framework for deployment
fastapi>=0.88.0
uvicorn>=0.20.0
requests>=2.28.0

# Serialization formats
safetensors>=0.3.0
onnx>=1.13.0
onnxruntime>=1.13.0
gguf>=0.17.1

# Monitoring and tracking
wandb>=0.13.0
mlflow>=2.0.0

# Progress tracking
tqdm>=4.64.0

# Optional dependencies (installed separately)
# bitsandbytes>=0.36.0  # For quantization
# deepspeed>=0.8.0      # For distributed training